Move helper functions from util.go to common.go
